error型を定義
